requirements.txt
pillow==11.0.0
numpy==2.0.2
opencv-contrib-python==4.10.0.84
scikit-learn==1.5.2
tensorflow==2.18.0
matplotlib==3.9.2